# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /course/weeks/{weekNumber}:
    get:
      operationId: getWeek
      tags: [course]
      summary: Get one week's lectures and assignment list
      parameters:
        - name: weekNumber
          in: path
          required: true
          schema: { type: integer, minimum: 1, maximum: 4 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Week"
  /course/lectures/{lectureId}:
    get:
      operationId: getLecture
      tags: [course]
      summary: Get the full lecture text for a topic
      parameters:
        - name: lectureId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
...
  /practice/sessions:
    post:
      operationId: startPracticeSession
      tags: [practice]
      summary: Start a new practice session (infinite, adaptive)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticeSessionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeSession"
  /practice/sessions/{sessionId}/next:
    post:
      operationId: nextPracticeProblem
      tags: [practice]
      summary: Generate the next problem (adaptive difficulty, de-novo)
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NextProblemInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeProblem"
  /practice/sessions/{sessionId}/grade:
    post:
      operationId: gradePracticeAnswer
      tags: [practice]
      summary: Grade a practice answer; returns correct/incorrect and explanation
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticeAnswerInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeGrade"
...
components:
  schemas:
...
    Lecture:
      type: object
      required: [id, topicId, title, weekNumber, body]
      properties:
        id: { type: integer }
        topicId: { type: integer }
        title: { type: string }
        weekNumber: { type: integer }
        body:
          type: string
          description: Short Markdown lecture text (the baseline / minimum-detail version). The frontend lets users select passages and send them to the tutor.
        bodyMedium:
          type: ["string", "null"]
          description: Medium-length version with more explanation and more examples. Null if not yet generated.
        bodyLong:
          type: ["string", "null"]
          description: Long version with the most explanation and the most examples. Null if not yet generated.
    Week:
      type: object
      required: [weekNumber, title, lectures, assignments]
      properties:
        weekNumber: { type: integer }
        title: { type: string }
        summary: { type: ["string", "null"] }
        lectures:
          type: array
          items:
            $ref: "#/components/schemas/LectureRef"
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/AssignmentSummary"
    LectureRef:
      type: object
      required: [id, title, topicId]
      properties:
        id: { type: integer }
        title: { type: string }
        topicId: { type: integer }
    AssignmentSummary:
      type: object
      required: [id, kind, title, weekNumber, problemCount, isTimed, status]
      properties:
        id: { type: integer }
        kind:
          type: string
          enum: [homework, test, midterm, final]
        title: { type: string }
        weekNumber: { type: integer }
        problemCount: { type: integer }
        isTimed: { type: boolean }
        timeLimitMinutes: { type: ["integer", "null"] }
        status:
          type: string
          enum: [not_started, in_progress, submitted]
        bestScore: { type: ["number", "null"] }
        lastAttemptId: { type: ["integer", "null"] }
        practiceRunCount: { type: ["integer", "null"] }
        bestPracticeScore: { type: ["number", "null"] }
        readinessLabel:
          type: ["string", "null"]
          enum: [not_ready, building, almost, ready, null]
...
    KeystrokeTrace:
      type: object
      required: [keystrokeCount, eraseCount, durationMs]
      properties:
        keystrokeCount: { type: integer }
        eraseCount: { type: integer }
        bulkInsertCount: { type: integer, default: 0 }
        longestBulkInsertChars: { type: integer, default: 0 }
        rewriteSegments: { type: integer, default: 0 }
        durationMs: { type: integer }
...
    PracticeSessionInput:
      type: object
      required: [tutorEnabled]
      properties:
        weekNumber: { type: ["integer", "null"] }
        topicId: { type: ["integer", "null"] }
        tutorEnabled: { type: boolean }
        focusOnWeaknesses: { type: boolean, default: true }
        initialDifficulty:
          type: ["number", "null"]
          description: Starting difficulty 1-5; if omitted defaults to 2.0
    PracticeSession:
      type: object
      required: [id, tutorEnabled, difficulty, weekNumber]
      properties:
        id: { type: integer }
        tutorEnabled: { type: boolean }
        difficulty: { type: number }
        weekNumber: { type: ["integer", "null"] }
        topicId: { type: ["integer", "null"] }
        focusOnWeaknesses: { type: boolean }
    NextProblemInput:
      type: object
      properties:
        request:
          type: ["string", "null"]
          description: Free-form user request, e.g. "give me harder probability problems"
        topicId: { type: ["integer", "null"] }
    PracticeProblem:
      type: object
      required: [id, prompt, topicId, difficulty]
      properties:
        id: { type: integer }
        prompt: { type: string }
        topicId: { type: integer }
        topicTitle: { type: string }
        difficulty: { type: number }
    PracticeAnswerInput:
      type: object
      required: [problemId, answer, trace]
      properties:
        problemId: { type: integer }
        answer: { type: string }
        trace:
          $ref: "#/components/schemas/KeystrokeTrace"
    PracticeGrade:
      type: object
      required: [problemId, correct, explanation, newDifficulty]
      properties:
        problemId: { type: integer }
        correct: { type: boolean }
        correctAnswer: { type: string }
        explanation: { type: string }
        newDifficulty: { type: number }
        tutorTip: { type: ["string", "null"] }
...