# lib/api-zod/src/generated/types/assignmentSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssignmentSummaryKind } from './assignmentSummaryKind';
import type { AssignmentSummaryReadinessLabel } from './assignmentSummaryReadinessLabel';
import type { AssignmentSummaryStatus } from './assignmentSummaryStatus';

export interface AssignmentSummary {
  id: number;
  kind: AssignmentSummaryKind;
  title: string;
  weekNumber: number;
  problemCount: number;
  isTimed: boolean;
  /** @nullable */
  timeLimitMinutes?: number | null;
  status: AssignmentSummaryStatus;
  /** @nullable */
  bestScore?: number | null;
  /** @nullable */
  lastAttemptId?: number | null;
  /** @nullable */
  practiceRunCount?: number | null;
  /** @nullable */
  bestPracticeScore?: number | null;
  /** @nullable */
  readinessLabel?: AssignmentSummaryReadinessLabel;
}